pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "visuallearn"
authors = [{name = "Noyonika Puram", email = "noyonikapuram@gmail.com"}]
description = "A Python library for visualizing ML model learning patterns and decision boundaries in real-time"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8", 
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Education",
]
keywords = ["machine learning", "visualization", "decision boundaries", "neural networks", "deep learning", "education"]
dependencies = [
    "matplotlib>=3.3.0",
    "numpy>=1.18.0", 
    "scikit-learn>=0.23.0",
    "imageio>=2.8.0",
    "Pillow>=7.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
pytorch = ["torch>=1.6.0"]
video = ["imageio[ffmpeg]"]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.10.0", 
    "black>=21.0.0",
    "flake8>=3.8.0",
    "mypy>=0.800",
]
all = ["torch>=1.6.0", "imageio[ffmpeg]"]

[project.urls]
Homepage = "https://github.com/noyonikap-glitch/visuallearn"
"Bug Reports" = "https://github.com/noyonikap-glitch/visuallearn/issues"
"Source" = "https://github.com/noyonikap-glitch/visuallearn"
"Documentation" = "https://github.com/noyonikap-glitch/visuallearn#readme"

[tool.setuptools_scm]
write_to = "visuallearn/_version.py"